+++ pkg/skaffold/deploy/kpt/kpt.go
@@ -44,6 +44,7 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/instrumentation"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/manifest"
+	kstatus "github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/status"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/log"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/output"
 	latestV1 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest/v1"
@@ -92,6 +93,7 @@ type Deployer struct {
 
 type Config interface {
 	kubectl.Config
+	kstatus.Config
 }
 
 // NewDeployer generates a new Deployer object contains the kptDeploy schema.
@@ -103,7 +105,7 @@ func NewDeployer(cfg Config, labels map[string]string, provider deploy.Component
 		accessor:           provider.Accessor.GetKubernetesAccessor(podSelector),
 		debugger:           provider.Debugger.GetKubernetesDebugger(podSelector),
 		logger:             provider.Logger.GetKubernetesLogger(podSelector),
-		statusMonitor:      provider.Monitor.GetKubernetesMonitor(),
+		statusMonitor:      provider.Monitor.GetKubernetesMonitor(cfg),
 		syncer:             provider.Syncer.GetKubernetesSyncer(podSelector),
 		insecureRegistries: cfg.GetInsecureRegistries(),
 		labels:             labels,
