+++ pkg/skaffold/deploy/kubectl/cli.go
@@ -32,6 +32,7 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/instrumentation"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubectl"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/manifest"
+	kstatus "github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/status"
 	latestV1 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest/v1"
 )
 
@@ -47,6 +48,7 @@ type CLI struct {
 
 type Config interface {
 	kubectl.Config
+	kstatus.Config
 	deploy.Config
 	ForceDeploy() bool
 	WaitForDeletions() config.WaitForDeletions
