+++ pkg/skaffold/kubernetes/context/context_test.go
@@ -196,7 +196,7 @@ func TestGetRestClientConfig(t *testing.T) {
 
 func TestUseKubeContext(t *testing.T) {
 	type invocation struct {
-		cliValue, yamlValue string
+		cliValue string
 	}
 	tests := []struct {
 		name        string
@@ -209,60 +209,10 @@ func TestUseKubeContext(t *testing.T) {
 			expected:    "",
 		},
 		{
-			name:        "yaml value when no CLI value is given",
-			invocations: []invocation{{yamlValue: "context2"}},
-			expected:    "context2",
-		},
-		{
-			name: "yaml value when no CLI value is given, first invocation persists",
-			invocations: []invocation{
-				{yamlValue: "context-first"},
-				{yamlValue: "context-second"},
-			},
-			expected: "context-first",
-		},
-		{
-			name:        "CLI value takes precedence",
-			invocations: []invocation{{cliValue: "context1", yamlValue: "context2"}},
-			expected:    "context1",
-		},
-		{
 			name: "first CLI value takes precedence",
 			invocations: []invocation{
 				{cliValue: "context-first"},
 				{cliValue: "context-second"},
-			},
-			expected: "context-first",
-		},
-		{
-			name: "mixed CLI value and yaml value - I",
-			invocations: []invocation{
-				{cliValue: "context-first"},
-				{yamlValue: "context-second"},
-			},
-			expected: "context-first",
-		},
-		{
-			name: "mixed CLI value and yaml value - II",
-			invocations: []invocation{
-				{yamlValue: "context-first"},
-				{cliValue: "context-second"},
-			},
-			expected: "context-first",
-		},
-		{
-			name: "mixed CLI value and yaml value - III",
-			invocations: []invocation{
-				{yamlValue: "context-first"},
-				{cliValue: "context-second", yamlValue: "context-third"},
-			},
-			expected: "context-first",
-		},
-		{
-			name: "mixed CLI value and yaml value - IV",
-			invocations: []invocation{
-				{cliValue: "context-first", yamlValue: "context-second"},
-				{cliValue: "context-third", yamlValue: "context-fourth"},
 			},
 			expected: "context-first",
 		},
@@ -272,7 +222,7 @@ func TestUseKubeContext(t *testing.T) {
 		testutil.Run(t, test.name, func(t *testutil.T) {
 			kubeContext = ""
 			for _, inv := range test.invocations {
-				ConfigureKubeConfig("", inv.cliValue, inv.yamlValue)
+				ConfigureKubeConfig("", inv.cliValue)
 			}
 
 			t.CheckDeepEqual(test.expected, kubeContext)
