+++ pkg/skaffold/runner/v1/new.go
@@ -18,9 +18,7 @@ package v1
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"strconv"
 
 	"github.com/sirupsen/logrus"
 
@@ -29,10 +27,6 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/cache"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/debug"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/helm"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/kpt"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/kubectl"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/kustomize"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/deploy/label"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/event"
 	eventV2 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/event/v2"
@@ -41,7 +35,6 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/instrumentation"
 	pkgkubectl "github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubectl"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes"
-	kstatus "github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes/status"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/log"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/runner/runcontext"
@@ -52,7 +45,6 @@ import (
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/tag"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/test"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/trigger"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
 )
 
 // NewForConfig returns a new SkaffoldRunner for a SkaffoldConfig
@@ -99,11 +91,11 @@ func NewForConfig(runCtx *runcontext.RunContext) (*SkaffoldRunner, error) {
 		Accessor: access.NewAccessorProvider(runCtx, labeller, kubectlCLI),
 		Debugger: debug.NewDebugProvider(runCtx),
 		Logger:   log.NewLogProvider(runCtx, kubectlCLI),
-		Monitor:  status.NewMonitorProvider(runCtx, labeller),
+		Monitor:  status.NewMonitorProvider(labeller),
 		Syncer:   sync.NewSyncProvider(runCtx, kubectlCLI),
 	}
 
-	deployer, podSelectors, err = getDeployer(runCtx, provider, labeller.Labels())
+	deployer, podSelectors, err = runner.GetDeployer(runCtx, provider, labeller.Labels())
 	if err != nil {
 		endTrace(instrumentation.TraceEndError(err))
 		return nil, fmt.Errorf("creating deployer: %w", err)
@@ -232,152 +224,4 @@ func getTester(cfg test.Config, isLocalImage func(imageName string) (bool, error
 	}
 
 	return tester, nil
-}
-
-/*
-The "default deployer" is used in `skaffold apply`, which uses a `kubectl` deployer to actuate resources
-on a cluster regardless of provided deployer configuration in the skaffold.yaml.
-The default deployer will honor a select set of deploy configuration from an existing skaffold.yaml:
-	- deploy.StatusCheckDeadlineSeconds
-	- deploy.Logs.Prefix
-	- deploy.Kubectl.Flags
-	- deploy.Kubectl.DefaultNamespace
-	- deploy.Kustomize.Flags
-	- deploy.Kustomize.DefaultNamespace
-For a multi-config project, we do not currently support resolving conflicts between differing sets of this deploy configuration.
-Therefore, in this function we do implicit validation of the provided configuration, and fail if any conflict cannot be resolved.
-*/
-func getDefaultDeployer(runCtx *runcontext.RunContext, provider deploy.ComponentProvider, labels map[string]string) (deploy.Deployer, kubernetes.ImageListMux, error) {
-	deployCfgs := runCtx.DeployConfigs()
-
-	var kFlags *latestV1.KubectlFlags
-	var logPrefix string
-	var defaultNamespace *string
-	var kubeContext string
-	statusCheckTimeout := -1
-
-	for _, d := range deployCfgs {
-		if d.KubeContext != "" {
-			if kubeContext != "" && kubeContext != d.KubeContext {
-				return nil, nil, errors.New("cannot resolve active Kubernetes context - multiple contexts configured in skaffold.yaml")
-			}
-			kubeContext = d.KubeContext
-		}
-		if d.StatusCheckDeadlineSeconds != 0 && d.StatusCheckDeadlineSeconds != int(kstatus.DefaultStatusCheckDeadline.Seconds()) {
-			if statusCheckTimeout != -1 && statusCheckTimeout != d.StatusCheckDeadlineSeconds {
-				return nil, nil, fmt.Errorf("found multiple status check timeouts in skaffold.yaml (not supported in `skaffold apply`): %d, %d", statusCheckTimeout, d.StatusCheckDeadlineSeconds)
-			}
-			statusCheckTimeout = d.StatusCheckDeadlineSeconds
-		}
-		if d.Logs.Prefix != "" {
-			if logPrefix != "" && logPrefix != d.Logs.Prefix {
-				return nil, nil, fmt.Errorf("found multiple log prefixes in skaffold.yaml (not supported in `skaffold apply`): %s, %s", logPrefix, d.Logs.Prefix)
-			}
-			logPrefix = d.Logs.Prefix
-		}
-		var currentDefaultNamespace *string
-		var currentKubectlFlags latestV1.KubectlFlags
-		if d.KubectlDeploy != nil {
-			currentDefaultNamespace = d.KubectlDeploy.DefaultNamespace
-			currentKubectlFlags = d.KubectlDeploy.Flags
-		}
-		if d.KustomizeDeploy != nil {
-			currentDefaultNamespace = d.KustomizeDeploy.DefaultNamespace
-			currentKubectlFlags = d.KustomizeDeploy.Flags
-		}
-		if kFlags == nil {
-			kFlags = &currentKubectlFlags
-		}
-		if err := validateKubectlFlags(kFlags, currentKubectlFlags); err != nil {
-			return nil, nil, err
-		}
-		if currentDefaultNamespace != nil {
-			if defaultNamespace != nil && *defaultNamespace != *currentDefaultNamespace {
-				return nil, nil, fmt.Errorf("found multiple namespaces in skaffold.yaml (not supported in `skaffold apply`): %s, %s", *defaultNamespace, *currentDefaultNamespace)
-			}
-			defaultNamespace = currentDefaultNamespace
-		}
-	}
-	if kFlags == nil {
-		kFlags = &latestV1.KubectlFlags{}
-	}
-	k := &latestV1.KubectlDeploy{
-		Flags:            *kFlags,
-		DefaultNamespace: defaultNamespace,
-	}
-	defaultDeployer, podSelector, err := kubectl.NewDeployer(runCtx, labels, provider, k)
-	if err != nil {
-		return nil, nil, fmt.Errorf("instantiating default kubectl deployer: %w", err)
-	}
-	return defaultDeployer, kubernetes.ImageListMux{podSelector}, nil
-}
-
-func validateKubectlFlags(flags *latestV1.KubectlFlags, additional latestV1.KubectlFlags) error {
-	errStr := "conflicting sets of kubectl deploy flags not supported in `skaffold apply` (flag: %s)"
-	if additional.DisableValidation != flags.DisableValidation {
-		return fmt.Errorf(errStr, strconv.FormatBool(additional.DisableValidation))
-	}
-	for _, flag := range additional.Apply {
-		if !util.StrSliceContains(flags.Apply, flag) {
-			return fmt.Errorf(errStr, flag)
-		}
-	}
-	for _, flag := range additional.Delete {
-		if !util.StrSliceContains(flags.Delete, flag) {
-			return fmt.Errorf(errStr, flag)
-		}
-	}
-	for _, flag := range additional.Global {
-		if !util.StrSliceContains(flags.Global, flag) {
-			return fmt.Errorf(errStr, flag)
-		}
-	}
-	return nil
-}
-
-func getDeployer(runCtx *runcontext.RunContext, provider deploy.ComponentProvider, labels map[string]string) (deploy.Deployer, kubernetes.ImageListMux, error) {
-	var podSelectors kubernetes.ImageListMux
-	if runCtx.Opts.Apply {
-		return getDefaultDeployer(runCtx, provider, labels)
-	}
-
-	deployerCfg := runCtx.Deployers()
-
-	var deployers deploy.DeployerMux
-	for _, d := range deployerCfg {
-		if d.HelmDeploy != nil {
-			h, podSelector, err := helm.NewDeployer(runCtx, labels, provider, d.HelmDeploy)
-			if err != nil {
-				return nil, nil, err
-			}
-			podSelectors = append(podSelectors, podSelector)
-			deployers = append(deployers, h)
-		}
-
-		if d.KptDeploy != nil {
-			deployer, podSelector := kpt.NewDeployer(runCtx, labels, provider, d.KptDeploy)
-			podSelectors = append(podSelectors, podSelector)
-			deployers = append(deployers, deployer)
-		}
-
-		if d.KubectlDeploy != nil {
-			deployer, podSelector, err := kubectl.NewDeployer(runCtx, labels, provider, d.KubectlDeploy)
-			if err != nil {
-				return nil, nil, err
-			}
-			podSelectors = append(podSelectors, podSelector)
-			deployers = append(deployers, deployer)
-		}
-
-		if d.KustomizeDeploy != nil {
-			deployer, podSelector, err := kustomize.NewDeployer(runCtx, labels, provider, d.KustomizeDeploy)
-			if err != nil {
-				return nil, nil, err
-			}
-			podSelectors = append(podSelectors, podSelector)
-			deployers = append(deployers, deployer)
-		}
-	}
-
-	return deployers, podSelectors, nil
 }
